Enemy 스케일 조절 후 프리팹 추가. Light 추가

+++ Assets/Scripts/BKH/HealthUI.cs
@@ -49,7 +49,7 @@ public class HealthUI : MonoBehaviour
         }
     }
 
-    void OnHealthChanged(int maxHealth, int currentHealth)
+    void OnHealthChanged(float maxHealth, float currentHealth)
     {
         if (ui != null)
         {
